// Data-Table-front-end/src/app/models/table.ts
export interface Table {
  Customer: string;
  JobNum: number;
  Item: string;
  Category: string;
  HpCopy: number;
  Tracking: string;
  Die: string;
  Plate: string;
  Varnish: string;
  Type: string;
  Mat1: string;
  Mat2: string;
  Width1: number;
  Width2: string;
  TaskQty: number;
  Press: number;
  Msf: number;
  Prodcode: string;
  Shipdate: string;
  Ordernotes: string;
}